requirements.txt
alembic==1.4.2
attrs==19.3.0
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
flask-expects-json==1.4.0
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
psycopg2==2.8.4
pyrsistent==0.15.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
